fix: extract startupGate to a side-effect-free module
